requirements.txt
pip
setuptools
wheel

python-dotenv
requests

mysql-connector-python
pg8000

psutil